Update QScheme data type to 16-bit unsigned integer.
This commit changes the QScheme data type to a 16-bit unsigned integer.
This is to have consistency with the BCQTensor in reading and saving quantization information, which eases calculating the offset.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/tensor/char_tensor.cpp
@@ -569,12 +569,12 @@ void CharTensor::copy(const void *buf) {
 }
 
 void CharTensor::save_quantization_info(std::ostream &file) {
-  checkedWrite(file, (char *)&qscheme, sizeof(uint8_t),
+  checkedWrite(file, (char *)&qscheme, sizeof(uint16_t),
                "[CharTensor::save] failed to write quantization information");
 }
 
 void CharTensor::read_quantization_info(std::ifstream &file) {
-  checkedRead(file, (char *)&qscheme, sizeof(uint8_t),
+  checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
               "[CharTensor::read] failed to read quantization information");
 }
